config: Using viper.IsSet() instead of custom code
Testing the presence of a key inside a viper configuration should be
done by the builtin call (IsSet) instead of a custom code.

The isEntryExists() function is kept as it wraps the group + item usage
into it. Pretty simple function but useful for the semantic.

Adding some tests to ensure a stable behavior over time.

Signed-off-by: Erwan Velu <erwan@redhat.com>